I removed a setMessageHandler:nil call from the AccessibilityBridge deconstructor.
Setting the message handler to nil is a hack to get leaking channels
to deallocate.  I already have a design doc for fixing that but it
hasn't been ratified yet.  I've also made the HandleEvent method a
function so that in the case that the channel is used after it is
already supposed to be deallocated it won't crash (since it doesn't
need a valid 'this').